requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
supervision>=0.18.0
mediapipe>=0.10.0
streamlit>=1.28.0
pygame>=2.5.0
pyserial>=3.5
Pillow>=10.0.0
plotly>=5.17.0
pyyaml>=6.0

